feat: set ID and description for the VIP 505
