+++ simulations/vip-505/bscmainnet.ts
@@ -4,7 +4,7 @@ import { ethers } from "hardhat";
 
 import { expectEvents } from "../../src/utils";
 import { forking, testVip } from "../../src/vip-framework";
-import vip501, {
+import vip505, {
   CERTIK,
   CERTIK_AMOUNT,
   FAIRYPROOF,
@@ -13,7 +13,7 @@ import vip501, {
   QUANTSTAMP_AMOUNT,
   USDC,
   USDT,
-} from "../../vips/vip-501/bscmainnet";
+} from "../../vips/vip-505/bscmainnet";
 import ERC20_ABI from "./abi/ERC20.json";
 import OMNICHAIN_PROPOSAL_SENDER_ABI from "./abi/OmnichainProposalSender.json";
 import VTREASURY_ABI from "./abi/VTreasury.json";
@@ -35,7 +35,7 @@ forking(50391406, async () => {
     prevUSDCBalanceOfQuantstamp = await usdc.balanceOf(QUANTSTAMP);
   });
 
-  testVip("VIP-501", await vip501(), {
+  testVip("VIP-505", await vip505(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(txResponse, [VTREASURY_ABI], ["WithdrawTreasuryBEP20"], [3]);
       await expectEvents(
