+++ simulations/vip-505/ethereum.ts
@@ -4,7 +4,7 @@ import { ethers } from "hardhat";
 import { expectEvents } from "src/utils";
 import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
 
-import { APXETH, DINERO, DINERO_AMOUNT, vip501 } from "../../vips/vip-501/bscmainnet";
+import { APXETH, DINERO, DINERO_AMOUNT, vip505 } from "../../vips/vip-505/bscmainnet";
 import ERC20_ABI from "./abi/ERC20.json";
 import VTREASURY_ABI from "./abi/VTreasuryEthereumAbi.json";
 
@@ -17,7 +17,7 @@ forking(22579756, async () => {
     prevAPXETHBalanceOfDinero = await apxETH.balanceOf(DINERO);
   });
 
-  testForkedNetworkVipCommands("VIP-501", await vip501(), {
+  testForkedNetworkVipCommands("VIP-505", await vip505(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(txResponse, [VTREASURY_ABI], ["WithdrawTreasuryToken"], [1]);
     },
